requirements.txt
discord.py[speed]
python-dotenv
APScheduler
asyncpg
pydantic
pydantic-settings
SQLAlchemy
psycopg2
aiocache
fastapi
uvicorn
markovify
# aiosqlite